+++ src/translations/components/TranslationFields/TranslationFields.tsx
@@ -15,6 +15,7 @@ import Skeleton from "@saleor/components/Skeleton";
 import TablePagination from "@saleor/components/TablePagination";
 import { buttonMessages } from "@saleor/intl";
 import { makeStyles } from "@saleor/macaw-ui";
+import { TranslationField } from "@saleor/translations/types";
 import { ListProps } from "@saleor/types";
 import classNames from "classnames";
 import React from "react";
@@ -23,14 +24,6 @@ import { FormattedMessage } from "react-intl";
 import TranslationFieldsLong from "./TranslationFieldsLong";
 import TranslationFieldsRich from "./TranslationFieldsRich";
 import TranslationFieldsShort from "./TranslationFieldsShort";
-
-export interface TranslationField {
-  displayName: string;
-  name: string;
-  translation: string;
-  type: "short" | "long" | "rich";
-  value: string;
-}
 
 type Pagination = Pick<
   ListProps,
@@ -47,7 +40,7 @@ export interface TranslationFieldsProps {
   pagination?: Pagination;
   onEdit: (field: string) => void;
   onDiscard: () => void;
-  onSubmit: (field: string, data: string | OutputData) => void;
+  onSubmit: (field: TranslationField, data: string | OutputData) => void;
 }
 
 const useStyles = makeStyles(
@@ -215,7 +208,7 @@ const TranslationFields: React.FC<TranslationFieldsProps> = props => {
                         initial={field.translation}
                         saveButtonState={saveButtonState}
                         onDiscard={onDiscard}
-                        onSubmit={data => onSubmit(field.name, data)}
+                        onSubmit={data => onSubmit(field, data)}
                       />
                     ) : field.type === "long" ? (
                       <TranslationFieldsLong
@@ -224,7 +217,7 @@ const TranslationFields: React.FC<TranslationFieldsProps> = props => {
                         initial={field.translation}
                         saveButtonState={saveButtonState}
                         onDiscard={onDiscard}
-                        onSubmit={data => onSubmit(field.name, data)}
+                        onSubmit={data => onSubmit(field, data)}
                       />
                     ) : (
                       <TranslationFieldsRich
@@ -233,7 +226,7 @@ const TranslationFields: React.FC<TranslationFieldsProps> = props => {
                         initial={field.translation}
                         saveButtonState={saveButtonState}
                         onDiscard={onDiscard}
-                        onSubmit={data => onSubmit(field.name, data)}
+                        onSubmit={data => onSubmit(field, data)}
                       />
                     )
                   ) : (