+++ src/translations/index.tsx
@@ -12,6 +12,9 @@ import {
   languageListPath,
   TranslatableEntities
 } from "./urls";
+import TranslationsAttributesComponent, {
+  TranslationsAttributesQueryParams
+} from "./views/TranslationsAttributes";
 import TranslationsCategoriesComponent, {
   TranslationsCategoriesQueryParams
 } from "./views/TranslationsCategories";
@@ -26,9 +29,6 @@ import TranslationsPagesComponent, {
 import TranslationsProductsComponent, {
   TranslationsProductsQueryParams
 } from "./views/TranslationsProducts";
-import TranslationsProductTypesComponent, {
-  TranslationsProductTypesQueryParams
-} from "./views/TranslationsProductTypes";
 import TranslationsSaleComponent, {
   TranslationsSalesQueryParams
 } from "./views/TranslationsSales";
@@ -155,16 +155,16 @@ const TranslationsPages: React.FC<TranslationsEntityRouteProps> = ({
     />
   );
 };
-const TranslationsProductTypes: React.FC<TranslationsEntityRouteProps> = ({
+const TranslationsAttributes: React.FC<TranslationsEntityRouteProps> = ({
   location,
   match
 }) => {
   const qs = parseQs(location.search.substr(1));
-  const params: TranslationsProductTypesQueryParams = {
+  const params: TranslationsAttributesQueryParams = {
     activeField: qs.activeField
   };
   return (
-    <TranslationsProductTypesComponent
+    <TranslationsAttributesComponent
       id={decodeURIComponent(match.params.id)}
       languageCode={LanguageCodeEnum[match.params.languageCode]}
       params={params}
@@ -263,10 +263,10 @@ const TranslationsRouter: React.FC = () => {
           exact
           path={languageEntityPath(
             ":languageCode",
-            TranslatableEntities.productTypes,
+            TranslatableEntities.attributes,
             ":id"
           )}
-          component={TranslationsProductTypes}
+          component={TranslationsAttributes}
         />
         <Route
           exact